Create dog_breed_CNN_pytorch.py
Part of the code from the second project of the Udacity Deep Learning Nanodegree. The task was to build a CNN from scratch which predicts the dog breed of a given image. The goal for the accuracy was 10%, this model achieved about 15%. For higher accuracy methods like transfer learning are used.